database update

+++ IFTTT/src/main/java/it/ifttt/repository/UserRepository.java
@@ -1,14 +1,14 @@
 package it.ifttt.repository;
 
+
 import org.springframework.data.repository.Repository;
 
-import it.ifttt.model.user.User;
+import it.ifttt.domain.User;
 
-public interface UserRepository extends Repository<User, Long> {
+public interface UserRepository extends Repository<User, Integer> {
 
-	User getById(long id);
-	
+	User getById(int id);
 	User findByUsername(String username);
-	
 	User save(User user);
+	  
 }
